requirements.txt
# Core dependencies
python-dotenv==1.0.0
telethon==1.34.0
openai==1.12.0
anthropic==0.18.1

# Trading APIs (choose based on your broker)
ccxt==4.2.25  # For crypto exchanges
alpaca-trade-api==3.0.2  # For stocks/forex (Alpaca)
python-binance==1.0.19  # For Binance specifically

# Data handling
pandas==2.2.0
numpy==1.26.3

# Async support
aiohttp==3.9.1
asyncio==3.4.3

# Logging and monitoring
loguru==0.7.2

# Database (optional, for trade history)
sqlalchemy==2.0.25
